requirements.txt
asn1crypto==0.22.0
bcrypt==3.1.3
cffi==1.10.0
cryptography==1.9
Django==1.11.2
django-mongodb-engine==0.6.0
django-registration-redux==1.6
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
djangotoolbox==1.8.0
idna==2.5
psycopg2==2.7.1
pycparser==2.17
PyJWT==1.5.0
pymongo==3.4.0
pytz==2017.2
six==1.10.0
